use 64bit random engine

+++ src/gcl.h
@@ -7,6 +7,7 @@
 
 #include <atomic>
 #include <chrono>
+#include <cstdint>
 #include <functional>
 #include <memory>
 #include <queue>
@@ -55,7 +56,7 @@ public:
 struct AsyncConfig
 {
     bool active = true;
-    std::size_t scheduler_random_seed = 0;
+    std::uint_fast64_t scheduler_random_seed = 0;
     bool processor_yields = true;
     bool scheduler_yields = true;
     std::chrono::microseconds inactive_processor_sleep_interval = std::chrono::microseconds{1000};
